+++ projects/core/src/clipboard/clipboard.service.ts
@@ -1,6 +1,6 @@
 import { Injectable, Optional, PLATFORM_ID, Inject } from "@angular/core";
 import { DOCUMENT, isPlatformBrowser } from "@angular/common";
-import { WINDOW } from "../window";
+import { WINDOW } from "../window/window.service";
 
 @Injectable()
 export class NgEzClipboardService{