+++ projects/core/src/in-viewport/in-viewport.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from "@angular/core";
 import { NgEzInViewportDirective } from "./in-viewport.directive";
-import { WINDOW_PROVIDERS } from "../window";
+import { WINDOW_PROVIDERS } from "../window/window.service";
 
 @NgModule({
     declarations: [NgEzInViewportDirective],