+++ projects/core/src/lazy-renderer/lazy-renderer.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from "@angular/core";
 import { NgEzLazyRendererComponent } from "./lazy-renderer.component";
-import { NgEzInViewportModule } from "../in-viewport";
+import { NgEzInViewportModule } from "../in-viewport/in-viewport.module";
 import { CommonModule } from "@angular/common";
 
 @NgModule({