+++ projects/core/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.1.0 (2019-01-02)
+### Features
+* Added **lazy-renderer** component.
+* Added **in-viewport** directive.
+
 # 1.0.2 (2018-12-28)
 ### Bug Fixes
 * **code-prettify:** should not throw warning on init.